@@ -1,29 +1,51 @@
 {
-  "name": "ngx-fullstory",
-  "version": "1.0.6",
-  "description": "A FullStory wrapper for Angular 6+",
-  "main": "index.js",
+  "name": "@nibo-ai/ngx-fullstory",
+  "version": "2.0.2",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "transpile": "ngc",
-    "package": "rollup -c",
-    "minify": "uglifyjs dist/bundles/ngx-fullstory.umd.js --screw-ie8 --compress --mangle --comments --output dist/bundles/ngx-fullstory.umd.min.js",
-    "build": "npm run transpile && npm run package && npm run minify"
+    "ng": "ng",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "build-prod": "ng build --prod",
+    "prepublish-lib": "npm run build-prod",
+    "publish-lib": "npm publish dist/ngx-fullstory"
   },
-  "keywords": [
-    "fullstory",
-    "angular"
-  ],
-  "author": "Marc Obaldo",
-  "license": "ISC",
+  "publishConfig": {
+    "access": "public"
+  },
   "devDependencies": {
-    "@angular/compiler": "^6.1.0",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/router": "^6.1.0",
-    "@angular/common": "^6.1.0",
-    "rollup": "^0.63.4",
-    "typescript": "^2.9.2",
-    "uglify-js": "^3.4.6"
+    "@angular/animations": "~9.0.1",
+    "@angular/common": "~9.0.1",
+    "@angular/compiler": "~9.0.1",
+    "@angular/core": "~9.0.1",
+    "@angular/forms": "~9.0.1",
+    "@angular/platform-browser": "~9.0.1",
+    "@angular/platform-browser-dynamic": "~9.0.1",
+    "@angular/router": "~9.0.1",
+    "rxjs": "~6.5.4",
+    "zone.js": "~0.10.2",
+    "@angular-devkit/build-angular": "~0.900.2",
+    "@angular-devkit/build-ng-packagr": "~0.900.2",
+    "@angular/cli": "~9.0.2",
+    "@angular/compiler-cli": "~9.0.1",
+    "@angular/language-service": "~9.0.1",
+    "@types/node": "^12.11.1",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^5.1.2",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "ng-packagr": "^9.0.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslib": "^1.10.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.7.5"
   }
 }
