@@ -15,7 +15,7 @@ import { FullstoryConfig } from './shared/fullstory-config';
   ]
 })
 export class FullstoryModule {
-  static forRoot(config: FullstoryConfig): ModuleWithProviders {
+  static forRoot(config: FullstoryConfig): ModuleWithProviders<FullstoryModule> {
     return {
       ngModule: FullstoryModule,
       providers: [
@@ -24,3 +24,5 @@ export class FullstoryModule {
     }
   }
 }
+
+export * from './fullstory/fullstory';
